requirements.txt

# Django and core dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3
python-decouple==3.8

# Authentication and security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7

# Database
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0

# Celery for background tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# WebSocket support
channels==4.0.0
channels-redis==4.1.0

# File storage and image processing
boto3==1.34.0
django-storages==1.14.2
Pillow==10.1.0
django-imagekit==5.0.0

# API documentation
drf-yasg==1.21.7
django-rest-swagger==2.2.0

# Payment integrations
requests==2.31.0
pycryptodome==3.19.0

# Monitoring and logging
sentry-sdk[django]==1.38.0
django-debug-toolbar==4.2.0

# Development and testing
django-seed==0.3.1
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.6.0
coverage==7.3.2

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
django-taggit==4.0.0
django-mptt==0.15.0
